python/scripts/scan_directory/pyproject.toml
[tool.poetry]
name = "scan-directory"
version = "1.0.0"
description = "Generador de árbol de directorios con soporte para exclusiones por YAML"
authors = ["Tu Nombre <tu.email@ejemplo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
